reorganzied cached loaders
