+++ src/app/padavali/(public)/page.tsx
@@ -15,7 +15,7 @@ import {
 export const dynamic = 'force-dynamic';
 
 async function buildListedPuzzlesInit(script: Awaited<ReturnType<typeof getCachedScript>>) {
-  const listed_puzzles = await CACHE.padavali_listed_puzzle_list.get(NO_CACHE_PARAMS);
+  const listed_puzzles = await CACHE.padavali.listed_puzzle_list.get(NO_CACHE_PARAMS);
 
   const puzzle_texts = listed_puzzles.flatMap((p) =>
     p.description ? [p.title, p.description] : [p.title]
@@ -39,8 +39,8 @@ async function buildListedPuzzlesInit(script: Awaited<ReturnType<typeof getCache
 
 export default async function Home() {
   const [current_schedule, next_schedule] = await Promise.all([
-    CACHE.padavali_current_schedule.get(NO_CACHE_PARAMS),
-    CACHE.padavali_next_schedule.get(NO_CACHE_PARAMS)
+    CACHE.padavali.current_schedule.get(NO_CACHE_PARAMS),
+    CACHE.padavali.next_schedule.get(NO_CACHE_PARAMS)
   ]);
 
   if (!current_schedule) {
